refactor(Client): create custom hook DateOfBirth
