+++ client/src/components/RegisterForm/components/DateOfBirthField.jsx
@@ -1,28 +1,10 @@
 import { getDays, getMonths, getYears } from '../../../utils'
+import { DateOfBirth } from '../hooks/DateOfBirth'
 import ContainerField from './ContainerField'
 import SelectionField from './SelectionField'
 
 const DateOfBirthField = ({ value, errors, onChange, onFocus, onBlur }) => {
-  const normalizeTargetInput = (e) => {
-    const { value: newValue, name: inputName } = e.target
-    return { name: 'date', value: { ...value, [inputName]: +newValue } }
-  }
-
-  const handleChange = (e) => {
-    const target = normalizeTargetInput(e)
-    onChange({ target })
-  }
-
-  const handleFocus = (e) => {
-    const target = normalizeTargetInput(e)
-    onFocus({ target })
-  }
-
-  const handleBlur = (e) => {
-    const target = normalizeTargetInput(e)
-    onBlur({ target })
-  }
-
+  const handlers = DateOfBirth({ value, errors, onChange, onFocus, onBlur })
   return (
     <ContainerField label="Fecha de nacimiento">
       <div className="register-form__wrapper-select">
@@ -31,27 +13,21 @@ const DateOfBirthField = ({ value, errors, onChange, onFocus, onBlur }) => {
           options={getDays()}
           value={value.day}
           errors={errors}
-          onChange={handleChange}
-          onFocus={handleFocus}
-          onBlur={handleBlur}
+          {...handlers}
         />
         <SelectionField
           name="month"
           options={getMonths()}
           value={value.month}
           errors={errors}
-          onChange={handleChange}
-          onFocus={handleFocus}
-          onBlur={handleBlur}
+          {...handlers}
         />
         <SelectionField
           name="year"
           options={getYears()}
           value={value.year}
           errors={errors}
-          onChange={handleChange}
-          onFocus={handleFocus}
-          onBlur={handleBlur}
+          {...handlers}
         />
       </div>
     </ContainerField>
